# Use SSE instruction to copy the data of structure