--- a/ext/oj/util.h
+++ b/ext/oj/util.h
@@ -5,6 +5,7 @@
 #define OJ_UTIL_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 typedef struct _timeInfo {
     int sec;
@@ -16,5 +17,6 @@ typedef struct _timeInfo {
 } * TimeInfo;
 
 extern void sec_as_time(int64_t secs, TimeInfo ti);
+extern void oj_memcpy(void *dest, const void *src, size_t len);
 
 #endif /* OJ_UTIL_H */