When compiled with gcc, the data is copied by the `rep movsq` instruction at the point where the structure is assigned and the 8 bytes are copied at a time.

```
       │    parse_info_init(&pi);
       │    pi.options   = oj_default_options;
  1.29 │      mov  oj_default_options@@Base-0x1748,%rsi
  0.12 │      mov  $0x2b,%ecx
       │      lea  0x10f8(%rsp),%rdi
 82.95 │      rep  movsq %ds:(%rsi),%es:(%rdi)
```

(The perf command shows the above result in my environment.)

The SSE instruction can be used to increase the size of the copy at a time.

−       | before | after  | result
--       | --     | --     | --
Oj.load  | 4.975M | 5.215M | 1.048x

### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 5.18.0-1-rt11-MANJARO
  - AMD Ryzen 7 5700G
  - gcc version 12.1.0
  - Ruby 3.1.2

### Before
```
Warming up --------------------------------------
             Oj.load   495.911k i/100ms
Calculating -------------------------------------
             Oj.load      4.975M (± 0.2%) i/s -     50.087M in  10.068529s
```

### After
```
Warming up --------------------------------------
             Oj.load   521.707k i/100ms
Calculating -------------------------------------
             Oj.load      5.215M (± 0.2%) i/s -     52.171M in  10.004304s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json = "42"

Benchmark.ips do |x|
  x.time = 10

  x.report('Oj.load') do |times|
    i = 0
    while i < times
      Oj.load(json)
      i += 1
    end
  end
end
```