@Watson1978 2022-07-15
Seems that original memcpy() is faster than my implementation...

```
Warming up --------------------------------------
             Oj.load   548.465k i/100ms
Calculating -------------------------------------
             Oj.load      5.476M (± 0.2%) i/s -     54.846M in  10.015495s
```