@Watson1978 2022-07-17
clang outputs highly optimized binaries, so it will slow down on clang env if applying this patch.

I'll investigate this for a while.