@Watson1978 2022-07-17
By the way, the article on https://danlark.org/2020/10/08/i-need-extra-c-c-performance-now-how/ is interesting